Fixes from PR#2
